fix: unify 16bit float naming and usage to bfloat16 (#895)
* unify 16bit float naming and usage to bfloat16


* fix data corruption during the renaming


---------

+++ collector/collect_oneccl_xpu.py
@@ -39,7 +39,7 @@ OP_NAME_TO_CCL = {
 
 # Mapping from our dtype names to oneCCL benchmark dtype names
 DTYPE_TO_CCL = {
-    "half": "float16",
+    "half": "bfloat16",
     "int8": "int8",
 }
